+++ crates/zizmor/src/audit/overprovisioned_secrets.rs
@@ -3,6 +3,7 @@ use std::ops::Deref;
 use github_actions_expressions::{
     Expr, SpannedExpr,
     call::{Call, Function},
+    op::BinExpr,
 };
 
 use crate::{
@@ -107,11 +108,11 @@ impl OverprovisionedSecrets {
                     _ => results.extend(ctx.parts.iter().flat_map(Self::secrets_expansions)),
                 }
             }
-            Expr::BinOp { lhs, op: _, rhs } => {
+            Expr::BinExpr(BinExpr { lhs, op: _, rhs }) => {
                 results.extend(Self::secrets_expansions(lhs));
                 results.extend(Self::secrets_expansions(rhs));
             }
-            Expr::UnOp { op: _, expr } => results.extend(Self::secrets_expansions(expr)),
+            Expr::UnExpr { op: _, expr } => results.extend(Self::secrets_expansions(expr)),
             _ => (),
         }