+++ crates/zizmor/src/audit/unsound_contains.rs
@@ -5,6 +5,7 @@ use github_actions_expressions::{
     call::{Call, Function},
     context::Context,
     literal::Literal,
+    op::BinExpr,
 };
 use github_actions_models::common::If;
 
@@ -106,13 +107,13 @@ impl UnsoundContains {
                 Box::new(exprs.iter().flat_map(Self::walk_tree_for_unsound_contains))
             }
             Expr::Index(expr) => Self::walk_tree_for_unsound_contains(expr),
-            Expr::BinOp { lhs, rhs, .. } => {
+            Expr::BinExpr(BinExpr { lhs, rhs, .. }) => {
                 let bc_lhs = Self::walk_tree_for_unsound_contains(lhs);
                 let bc_rhs = Self::walk_tree_for_unsound_contains(rhs);
 
                 Box::new(bc_lhs.chain(bc_rhs))
             }
-            Expr::UnOp { expr, .. } => Self::walk_tree_for_unsound_contains(expr),
+            Expr::UnExpr { expr, .. } => Self::walk_tree_for_unsound_contains(expr),
             _ => Box::new(std::iter::empty()),
         }
     }